+++ internal/storage/sqlite/categories.go
@@ -8,13 +8,14 @@ import (
 	"expense-tracker-api/internal/storage"
 
 	"github.com/google/uuid"
+	"github.com/mattn/go-sqlite3"
 )
 
 func (s *Storage) CreateCategory(params storage.CreateCategoryParams) (*storage.Category, error) {
 	const op = "storage.sqlite.CreateCategory"
 
 	stmt, err := s.db.Prepare(
-		`INSERT INTO categories (id, name, type, icon, color, user_id) VALUES (?, ?, ?, ?, ?, ?) RETURNING id, name, type, icon, color, created_at, updated_at`,
+		`INSERT INTO categories (id, user_id, name, type, icon, color) VALUES (?, ?, ?, ?, ?, ?) RETURNING id, user_id, name, type, icon, color, created_at, updated_at`,
 	)
 	if err != nil {
 		return nil, fmt.Errorf("%s: %w", op, err)
@@ -23,9 +24,13 @@ func (s *Storage) CreateCategory(params storage.CreateCategoryParams) (*storage.
 
 	id := uuid.NewString()
 	var category storage.Category
-	err = stmt.QueryRow(id, params.Name, params.Type, params.Icon, params.Color, params.UserId).
-		Scan(&category.Id, &category.Name, &category.Type, &category.Icon, &category.Color, &category.CreatedAt, &category.UpdatedAt)
+	err = stmt.QueryRow(id, params.UserID, params.Name, params.Type, params.Icon, params.Color).
+		Scan(&category.ID, &category.UserID, &category.Name, &category.Type, &category.Icon, &category.Color, &category.CreatedAt, &category.UpdatedAt)
 	if err != nil {
+		var sqliteErr sqlite3.Error
+		if errors.As(err, &sqliteErr) && sqliteErr.ExtendedCode == sqlite3.ErrConstraintUnique {
+			return nil, fmt.Errorf("%s: %w", op, storage.ErrCategoryAlreadyExists)
+		}
 		return nil, fmt.Errorf("%s: %w", op, err)
 	}
 
@@ -33,6 +38,7 @@ func (s *Storage) CreateCategory(params storage.CreateCategoryParams) (*storage.
 }
 
 func (s *Storage) UpdateCategory(
+	userID string,
 	id string,
 	params storage.UpdateCategoryParams,
 ) (*storage.Category, error) {
@@ -46,15 +52,16 @@ func (s *Storage) UpdateCategory(
 		build(", ")
 
 	args = append(args, id)
+	args = append(args, userID)
 
 	query := fmt.Sprintf(
-		`UPDATE categories SET %s WHERE id = ? RETURNING id, name, type, icon, color, created_at, updated_at`,
+		`UPDATE categories SET %s WHERE id = ? AND user_id = ? RETURNING id, user_id, name, type, icon, color, created_at, updated_at`,
 		setParts,
 	)
 
 	var category storage.Category
 	err := s.db.QueryRow(query, args...).
-		Scan(&category.Id, &category.Name, &category.Type, &category.Icon, &category.Color, &category.CreatedAt, &category.UpdatedAt)
+		Scan(&category.ID, &category.UserID, &category.Name, &category.Type, &category.Icon, &category.Color, &category.CreatedAt, &category.UpdatedAt)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, fmt.Errorf("%s: %w", op, storage.ErrCategoryNotFound)
@@ -65,18 +72,18 @@ func (s *Storage) UpdateCategory(
 	return &category, nil
 }
 
-func (s *Storage) DeleteCategory(id string) error {
+func (s *Storage) DeleteCategory(userID string, id string) error {
 	const op = "storage.sqlite.DeleteCategory"
 
 	stmt, err := s.db.Prepare(
-		`DELETE FROM categories WHERE id = ?`,
+		`DELETE FROM categories WHERE id = ? AND user_id = ?`,
 	)
 	if err != nil {
 		return fmt.Errorf("%s: %w", op, err)
 	}
 	defer stmt.Close()
 
-	res, err := stmt.Exec(id)
+	res, err := stmt.Exec(id, userID)
 	if err != nil {
 		if isFKViolationError(err) {
 			return fmt.Errorf("%s: %w", op, storage.ErrCategoryHasTransactions)
@@ -95,11 +102,11 @@ func (s *Storage) DeleteCategory(id string) error {
 	return nil
 }
 
-func (s *Storage) GetCategory(id string) (*storage.Category, error) {
+func (s *Storage) GetCategory(userID string, id string) (*storage.Category, error) {
 	const op = "storage.sqlite.GetCategory"
 
 	stmt, err := s.db.Prepare(
-		`SELECT id, name, type, icon, color, created_at, updated_at FROM categories WHERE id = ?`,
+		`SELECT id, user_id, name, type, icon, color, created_at, updated_at FROM categories WHERE id = ? AND user_id = ?`,
 	)
 	if err != nil {
 		return nil, fmt.Errorf("%s: %w", op, err)
@@ -107,8 +114,8 @@ func (s *Storage) GetCategory(id string) (*storage.Category, error) {
 	defer stmt.Close()
 
 	var category storage.Category
-	err = stmt.QueryRow(id).
-		Scan(&category.Id, &category.Name, &category.Type, &category.Icon, &category.Color, &category.CreatedAt, &category.UpdatedAt)
+	err = stmt.QueryRow(id, userID).
+		Scan(&category.ID, &category.UserID, &category.Name, &category.Type, &category.Icon, &category.Color, &category.CreatedAt, &category.UpdatedAt)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, fmt.Errorf("%s: %w", op, storage.ErrCategoryNotFound)
@@ -119,12 +126,17 @@ func (s *Storage) GetCategory(id string) (*storage.Category, error) {
 	return &category, nil
 }
 
-func (s *Storage) GetCategories(params storage.GetCategoriesParams) ([]storage.Category, error) {
+func (s *Storage) GetCategories(
+	userID string,
+	params storage.GetCategoriesParams,
+) ([]storage.Category, error) {
 	const op = "storage.sqlite.GetCategories"
 
-	query := `SELECT id, name, type, icon, color, created_at, updated_at FROM categories`
+	query := `SELECT id, user_id, name, type, icon, color, created_at, updated_at FROM categories`
 	whereParts, args := newWhereBuilder().
-		addString("type", params.Type).build(" AND ")
+		addString("type", params.Type).
+		addString("user_id", &userID).
+		build(" AND ")
 
 	if len(whereParts) > 0 {
 		query = fmt.Sprintf(`%s WHERE %s`, query, whereParts)
@@ -140,7 +152,8 @@ func (s *Storage) GetCategories(params storage.GetCategoriesParams) ([]storage.C
 	for rows.Next() {
 		category := storage.Category{}
 		err := rows.Scan(
-			&category.Id,
+			&category.ID,
+			&category.UserID,
 			&category.Name,
 			&category.Type,
 			&category.Icon,
